requirements.txt
flask
googletrans
youtube_transcript_api
moviepy
ffmpeg
pytube
biopython
audalign